Fix JSON format of time events
