+++ src/Trace/TimestampTrait.php
@@ -1,0 +1,49 @@
+<?php
+/**
+ * Copyright 2017 Google Inc. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+namespace Google\Cloud\Trace;
+
+/**
+ * Trait which provides helper methods for formatting timestamps.
+ */
+trait TimestampTrait
+{
+    /**
+     * Returns a "Zulu" formatted string representing the provided \DateTime.
+     *
+     * @param  \DateTimeInterface|int|float|string $when [optional] The end time of this span.
+     *         **Defaults to** now. If provided as a string, it must be in "Zulu" format.
+     *         If provided as an int or float, it is expected to be a Unix timestamp.
+     * @return string
+     */
+    private function formatDate($when = null)
+    {
+        if (is_string($when)) {
+            return $when;
+        } elseif (!$when) {
+            list($usec, $sec) = explode(' ', microtime());
+            $micro = sprintf("%06d", $usec * 1000000);
+            $when = new \DateTime(date('Y-m-d H:i:s.' . $micro));
+        } elseif (is_numeric($when)) {
+            // Expect that this is a timestamp
+            $micro = sprintf("%06d", ($when - floor($when)) * 1000000);
+            $when = new \DateTime(date('Y-m-d H:i:s.'. $micro, (int) $when));
+        }
+        $when->setTimezone(new \DateTimeZone('UTC'));
+        return $when->format('Y-m-d\TH:i:s.u000\Z');
+    }
+}
